small fix suggested by Hubert
On 14.09.2013, at 12:52, Hubert Wagner <hub.wag@gmail.com> wrote:

Hi Uli,

I am using PHAT for some experiments and it didn't compile under VC++ 2010. The problem was that std::back_inserter was not found in a number of cases. The solution is to  #include <iterator> in helpers/misc.h.

Best,
Hubert